+++ chain/messagepool/selection.go
@@ -11,7 +11,9 @@ import (
 
 	"github.com/filecoin-project/go-address"
 	tbig "github.com/filecoin-project/go-state-types/big"
+	"github.com/filecoin-project/specs-actors/v3/actors/builtin"
 
+	"github.com/filecoin-project/lotus/api"
 	"github.com/filecoin-project/lotus/build"
 	"github.com/filecoin-project/lotus/chain/messagepool/gasguess"
 	"github.com/filecoin-project/lotus/chain/types"
@@ -698,6 +700,17 @@ func (*MessagePool) getGasPerf(gasReward *big.Int, gasLimit int64) float64 {
 	return r
 }
 
+func isMessageMute(m *types.Message, ts *types.TipSet) bool {
+	if api.RunningNodeType != api.NodeFull || ts.Height() > build.UpgradeActorsV4Height {
+		return false
+	}
+
+	if m.To == builtin.StoragePowerActorAddr {
+		return m.Method == builtin.MethodsPower.CreateMiner
+	}
+	return false
+}
+
 func (mp *MessagePool) createMessageChains(actor address.Address, mset map[uint64]*types.SignedMessage, baseFee types.BigInt, ts *types.TipSet) []*msgChain {
 	// collect all messages
 	msgs := make([]*types.SignedMessage, 0, len(mset))
@@ -758,6 +771,11 @@ func (mp *MessagePool) createMessageChains(actor address.Address, mset map[uint6
 		if balance.Cmp(required) < 0 {
 			break
 		}
+
+		if isMessageMute(&m.Message, ts) {
+			break
+		}
+
 		balance = new(big.Int).Sub(balance, required)
 
 		value := m.Message.Value.Int