+++ chain/types/state.go
@@ -11,8 +11,10 @@ const (
 	StateTreeVersion0 StateTreeVersion = iota
 	// StateTreeVersion1 corresponds to actors v2
 	StateTreeVersion1
-	// StateTreeVersion2 corresponds to actors >= v3.
+	// StateTreeVersion2 corresponds to actors v3.
 	StateTreeVersion2
+	// StateTreeVersion3 corresponds to actors >= v4.
+	StateTreeVersion3
 )
 
 type StateRoot struct {