+++ CHANGELOG.md
@@ -67,6 +67,14 @@ This is an optional Lotus release that introduces various improvements to the se
 - fix 4857: show help for set-addrs (https://github.com/filecoin-project/lotus/pull/5943)
 - fix health report (https://github.com/filecoin-project/lotus/pull/6011)
 
+# 1.8.0 / 2021-04-05
+
+This is a mandatory release of Lotus that upgrades the network to version 12, which introduces various performance improvements to the cron processing of the power actor. The network will upgrade at height 712320, which is 2021-04-29T06:00:00Z.
+
+## Changes
+
+- v4 specs-actors integration, nv12 migration (https://github.com/filecoin-project/lotus/pull/6116)
+
 # 1.6.0 / 2021-04-05
 
 This is a mandatory release of Lotus that upgrades the network to version 11, which implements [FIP-0014](https://github.com/filecoin-project/FIPs/blob/master/FIPS/fip-0014.md). The network will upgrade at height 665280, which is 2021-04-12T22:00:00Z.